Finished 04-removeFromArray
Using the same concept as the solution, you are
given an array and another array of arguments
to "remove" or exclude from the given array. By
continuously checking if the given number is
on that removal list, we can push all the
elements that aren't on that list of numbers to
remove to create an array without those numbers.